Camera-relative world scrolling

+++ src/DummyLib/states/UITest2.cpp
@@ -83,23 +83,23 @@ void UITest2::OnPostloadScene(Sys::JsObjectP &js_scene) {
 
     BaseState::OnPostloadScene(js_scene);
 
-    Ren::Vec3f view_origin, view_dir = Ren::Vec3f{0, 0, 1};
+    Ren::Vec3d view_origin, view_dir = Ren::Vec3d{0, 0, 1};
     float view_fov = 45, min_exposure = -1000, max_exposure = 1000;
 
     if (js_scene.Has("camera")) {
         const Sys::JsObjectP &js_cam = js_scene.at("camera").as_obj();
         if (js_cam.Has("view_origin")) {
             const Sys::JsArrayP &js_orig = js_cam.at("view_origin").as_arr();
-            view_origin[0] = float(js_orig.at(0).as_num().val);
-            view_origin[1] = float(js_orig.at(1).as_num().val);
-            view_origin[2] = float(js_orig.at(2).as_num().val);
+            view_origin[0] = js_orig.at(0).as_num().val;
+            view_origin[1] = js_orig.at(1).as_num().val;
+            view_origin[2] = js_orig.at(2).as_num().val;
         }
 
         if (js_cam.Has("view_dir")) {
             const Sys::JsArrayP &js_dir = js_cam.at("view_dir").as_arr();
-            view_dir[0] = float(js_dir.at(0).as_num().val);
-            view_dir[1] = float(js_dir.at(1).as_num().val);
-            view_dir[2] = float(js_dir.at(2).as_num().val);
+            view_dir[0] = js_dir.at(0).as_num().val;
+            view_dir[1] = js_dir.at(1).as_num().val;
+            view_dir[2] = js_dir.at(2).as_num().val;
         }
 
         /*if (js_cam.Has("fwd_speed")) {
